+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 #### next release (8.3.7)
 
+- Refactor 3D & 2D zoom code path.
+- Improve zooming to clamped items in 3D.
 - [The next improvement]
 
 #### 8.3.6 - 2023-10-03